requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-jwt==1.18.0
generics==3.3.0
Pillow==8.1.2
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
